# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Heymaty - OpenAPI 3.0",
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://heymaty.com/"
    }
  ],
...
  "paths": {
...
    "/bundles/": {
      "get": {
        "tags": [
          "bundles"
        ],
        "summary": "Get all my bundles",
        "description": "Get all my bundles",
        "operationId": "Get all my bundles",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundlesResponse"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      },
      "post": {
        "tags": [
          "bundles"
        ],
        "summary": "Create a new bundle",
        "description": "Create a new bundle for a user",
        "operationId": "createbundles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "qty": {
                    "type": "integer",
                    "description": "Quantity of messages in the bundle"
                  },
                  "price": {
                    "type": "number",
                    "description": "Price of the bundle"
                  },
                  "isActive": {
                    "type": "boolean",
                    "description": "Status of the bundle (active or not)"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the bundle"
                  },
                  "type": {
                    "type": "string",
                    "description": "Type of the bundle (e.g., \"limited\")"
                  },
                  "amount": {
                    "type": "integer",
                    "description": "Amount of packages for a limited bundle"
                  }
                },
                "required": [
                  "qty",
                  "price"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "newBundle": {
                      "$ref": "#/components/schemas/Bundle"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      }
    },
    "/bundles/{bundleId}": {
      "get": {
        "tags": [
          "bundles"
        ],
        "summary": "Get a bundle",
        "description": "Get a bundle",
        "operationId": "Get a bundle",
        "parameters": [
          {
            "name": "bundleId",
            "in": "path",
            "description": "The type of callback to be processed.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bundle": {
                      "$ref": "#/components/schemas/Bundle"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "bundles"
        ],
        "summary": "Delete a bundle",
        "description": "Delete a bundle",
        "operationId": "Delete a bundle",
        "parameters": [
          {
            "name": "bundleId",
            "in": "path",
            "description": "The type of callback to be processed.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      }
    },
    "/bundles/model/{modelId}": {
      "get": {
        "tags": [
          "bundles"
        ],
        "summary": "Get all active bundles of model",
        "description": "Get all active bundles of model",
        "operationId": "Get all active bundles of model",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "description": "The type of callback to be processed.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundlesResponse"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      }
    },
...
    "/purchases/bundle/": {
      "post": {
        "tags": [
          "purchases"
        ],
        "summary": "Buy a bundle",
        "description": "Buy a bundle",
        "operationId": "Buy a bundle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bundleId": {
                    "type": "string"
                  }
                },
                "required": [
                  "bundleid"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "savedUser": {
                      "$ref": "#/components/schemas/User"
                    },
                    "message": {
                      "type": "string",
                      "example": "Пакет успешно куплен"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "user"
        ]
      }
    },
...
    "/tariff/": {
      "post": {
        "tags": [
          "tariff"
        ],
        "summary": "Create tariff",
        "description": "Create tariff",
        "operationId": "createTariff",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "description"
                  },
                  "rate": {
                    "type": "string",
                    "description": "rate"
                  }
                },
                "required": [
                  "description",
                  "rate"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tariff"
                }
              }
            }
          },
          "400": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      }
    },
    "/tariff/{id}": {
      "patch": {
        "tags": [
          "tariff"
        ],
        "summary": "Update tariff",
        "description": "Update tariff",
        "operationId": "updateTariff",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "description"
                  },
                  "rate": {
                    "type": "string",
                    "description": "rate"
                  }
                },
                "required": [
                  "description",
                  "rate"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Service": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the service."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the service was created.",
            "default": "The current date-time"
          }
        }
      },
...
      "Tariff": {
        "type": "object",
        "required": [
          "user",
          "description",
          "rate"
        ],
        "properties": {
          "user": {
            "type": "string",
            "description": "Reference to User model.",
            "format": "objectId"
          },
          "description": {
            "type": "string",
            "description": "Description of the tariff."
          },
          "rate": {
            "type": "number",
            "format": "float",
            "description": "Rate of the tariff."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the tariff was created."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the tariff was last updated."
          }
        }
      },
...
      "ServerErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "msg": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "errors"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "errors"
        ]
      },
...
      "Bundle": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "description": "Status of the bundle (active or not)"
          },
          "type": {
            "type": "string",
            "enum": [
              "limited",
              "unlimited"
            ],
            "description": "Type of the bundle"
          },
          "amount": {
            "type": "integer",
            "description": "Amount of packages (relevant for limited bundles)"
          },
          "payments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of payments related to the bundle"
          },
          "_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the bundle"
          },
          "user": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the user"
          },
          "qty": {
            "type": "integer",
            "description": "Quantity of messages in the bundle"
          },
          "price": {
            "type": "number",
            "description": "Price of the bundle"
          },
          "description": {
            "type": "string",
            "description": "Description of the bundle"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the bundle was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the bundle was last updated"
          },
          "__v": {
            "type": "integer",
            "description": "Version field for Mongoose schema"
          }
        }
      },
      "BundlesResponse": {
        "type": "object",
        "properties": {
          "bundles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bundle"
            }
          },
          "total": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 30,
            "description": "Name of the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user"
          },
          "emailConfirmed": {
            "type": "boolean",
            "default": false
          },
          "emailConfirmationCode": {
            "type": "string"
          },
          "emailConfirmationValidUntil": {
            "type": "string",
            "format": "date-time"
          },
          "emailConfirmationRequestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "girl",
              "admin",
              "moderator"
            ],
            "default": "user"
          },
          "verification": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "subscribers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subscribed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subscriptionEnabled": {
            "type": "boolean",
            "default": false
          },
          "profileNum": {
            "type": "integer",
            "format": "int64"
          },
          "profileUrl": {
            "type": "string"
          },
          "profileGallery": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "verified": {
            "type": "boolean",
            "default": false
          },
          "balance": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "unconfirmedBalance": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "inBTCAddress": {
            "type": "string"
          },
          "outBTCAddress": {
            "type": "string"
          },
          "paymentCode": {
            "type": "string"
          },
          "invoice": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "default": 18
          },
          "phone": {
            "type": "string"
          },
          "country": {
            "type": "object"
          },
          "state": {
            "type": "object"
          },
          "city": {
            "type": "object"
          },
          "profile": {
            "type": "object"
          },
          "cover": {
            "type": "object"
          },
          "nationalId": {
            "type": "object"
          },
          "pricePerDay": {
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          "pricePerMessage": {
            "type": "integer",
            "format": "int32",
            "default": 9999
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            }
          },
          "urlTelegram": {
            "type": "string",
            "format": "uri"
          },
          "urlWhatsApp": {
            "type": "string",
            "format": "uri"
          },
          "loginAttempts": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "blockExpires": {
            "type": "string",
            "format": "date-time",
            "default": "2023-09-14T00:00:00Z"
          },
          "googleId": {
            "type": "string"
          },
          "messageBundles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "user": {
                  "$ref": "#/components/schemas/User"
                },
                "messages": {
                  "type": "number"
                }
              }
            }
          },
          "personalFirstName": {
            "type": "string"
          },
          "personalSecondName": {
            "type": "string"
          },
          "personalFamilyName": {
            "type": "string"
          },
          "personalDayOfBirth": {
            "type": "integer",
            "format": "int32"
          },
          "personalMonthOfBirth": {
            "type": "integer",
            "format": "int32"
          },
          "personalYearOfBirth": {
            "type": "integer",
            "format": "int32"
          },
          "personalSex": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ]
          },
          "personalCountry": {
            "type": "string"
          },
          "personalCity": {
            "type": "string"
          },
          "personalStreet": {
            "type": "string"
          },
          "personalBuilding": {
            "type": "string"
          },
          "personalAppartment": {
            "type": "string"
          },
          "personalZIP": {
            "type": "string"
          },
          "personalPassport": {
            "type": "object"
          },
          "personalIDCardFront": {
            "type": "object"
          },
          "personalIDCardBack": {
            "type": "object"
          },
          "personalDriverLicenseFront": {
            "type": "object"
          },
          "personalDriverLicenseBack": {
            "type": "object"
          },
          "personalIDExpirationDay": {
            "type": "integer",
            "format": "int32"
          },
          "personalIDExpirationMonth": {
            "type": "integer",
            "format": "int32"
          },
          "personalIDExpirationYear": {
            "type": "integer",
            "format": "int32"
          },
          "personalIDHasNoExpiration": {
            "type": "boolean",
            "default": false
          },
          "personalSelfie": {
            "type": "object"
          }
        }
      },
      "Language": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the service."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "2023-09-14T00:00:00Z"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}